Refactor cyral_integration_*
