+++ cyral/internal/integration/awsiam/constants.go
@@ -1,0 +1,9 @@
+package awsiam
+
+const (
+	resourceName = "cyral_integration_aws_iam"
+
+	AWSIAMIntegrationNameKey        = "name"
+	AWSIAMIntegratioNDescriptionKey = "description"
+	AWSIAMIntegrationARNsKey        = "role_arns"
+)
