+++ cyral/internal/integration/confextension/constants.go
@@ -1,0 +1,9 @@
+package confextension
+
+const (
+	authorizationPurpose = "authorization"
+	builtinCategory      = "builtin"
+
+	PagerDutyTemplateType = "pagerduty"
+	DuoMFATemplateType    = "duoMfa"
+)
