+++ cyral/internal/integration/confextension/mfaduo/constants.go
@@ -1,0 +1,5 @@
+package mfaduo
+
+const (
+	resourceName = "cyral_integration_mfa_duo"
+)
