+++ cyral/internal/integration/confextension/pagerduty/constants.go
@@ -1,0 +1,5 @@
+package pagerduty
+
+const (
+	resourceName = "cyral_integration_pager_duty"
+)
