+++ cyral/internal/integration/idpsaml/constants.go
@@ -1,0 +1,7 @@
+package idpsaml
+
+const (
+	resourceName          = "cyral_integration_idp_saml"
+	resourceSamlDraftName = "cyral_integration_idp_saml_draft"
+	dataSourceName        = "cyral_integration_idp_saml"
+)
