+++ cyral/internal/integration/idpsaml/draft/constants.go
@@ -1,0 +1,5 @@
+package idpsaml_draft
+
+const (
+	resourceName = "cyral_integration_idp_saml_draft"
+)
