+++ cyral/internal/integration/logging/resource_test.go
